ARM/mm: Normalise page counts to unsigned long for ARM
 - xen/arch/arm/include/asm/p2m.h: guest_physmap_add_pages() took
   unsigned int nr_pages, which silently truncated unsigned long
   counts from common callers (notably static-shmem.c, which passes
   PFN_DOWN(psize) from an unsigned long nr_pages).  Widened to
   unsigned long.

 - xen/common/device-tree/static-memory.c: local nr_pages widened
   from unsigned int to unsigned long to match the caller chain.

 - device-tree: Replace UINT_MAX with ULONG_MAX in the d->max_pages
   overflow guard for symmetry with the widened field.

Already correct:

 - struct paging_domain.p2m_total_pages is unsigned long
   (xen/arch/arm/include/asm/domain.h).

 - p2m_set_allocation() takes unsigned long pages
   (xen/arch/arm/mmu/p2m.c).

 - arch_get_paging_mempool_size() uses (uint64_t)... << PAGE_SHIFT,

 - p2m_insert_mapping(), map_regions_p2mt(), unmap_regions_p2mt(),
   map_mmio_regions(), unmap_mmio_regions() all already take
   unsigned long nr.

+++ xen/common/device-tree/static-memory.c
@@ -11,7 +11,7 @@ static bool __init append_static_memory_to_bank(struct domain *d,
                                                 paddr_t size)
 {
     int res;
-    unsigned int nr_pages = PFN_DOWN(size);
+    unsigned long nr_pages = PFN_DOWN(size);
     gfn_t sgfn;
 
     /*
